Each codegen class in its own file
